Ship tiptap as rails rte (remove quill/lexxy)
