+++ playbook-website/config/menu.yml
@@ -519,7 +519,6 @@ kits:
     platforms: *1
     description: Full-featured text editor with formatting options.
     status: stable
-    platforms_status: { rails: deprecated }
     icons_used: true
     react_rendered: true
     enhanced_element_used: false
