--- playbook/app/pb_kits/playbook/pb_rich_text_editor/_lexxy_styles.scss
@@ -1,69 +1,0 @@
-// Playbook-themed overrides for Lexxy POC (toolbar + icons).
-// Scoped to [data-pb-rte-lexxy] so default Lexxy styling is unchanged elsewhere.
-
-@import "../tokens/border_radius";
-@import "../tokens/colors";
-@import "../tokens/spacing";
-@import "../tokens/transition";
-
-[data-pb-rte-lexxy] {
-  // Remove Lexxy’s outer wrapper border and padding so only toolbar + content form one card
-  lexxy-editor,
-  lexxy-editor.lexxy-content {
-    border: none;
-    border-radius: 0;
-    box-shadow: none;
-    padding: 0;
-  }
-
-  lexxy-toolbar {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    gap: $space_xs;
-    background: $white;
-    border: 1px solid $input_border_default;
-    border-bottom: none;
-    border-radius: $border_rad_heaviest $border_rad_heaviest 0 0;
-    padding: $space_xs;
-    @include transition_default;
-  }
-
-  .lexxy-editor__content {
-    border: 1px solid $input_border_default;
-    border-radius: 0 0 $border_rad_heaviest $border_rad_heaviest;
-  }
-
-  .lexxy-editor__toolbar-button {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background: transparent;
-    border: none;
-    border-radius: $border_rad_heaviest;
-    color: $text_lt_light;
-    width: $space_xl + 2;
-    height: $space_xl + 2;
-    cursor: pointer;
-    @include transition_default;
-
-    &:hover:not([disabled]) {
-      background-color: $neutral_subtle;
-    }
-
-    &[aria-pressed="true"] {
-      color: $primary;
-      background-color: $bg_light;
-    }
-
-    &:focus-visible {
-      outline: none;
-    }
-  }
-
-  .lexxy-editor__toolbar-button svg {
-    width: 1.25rem;
-    height: 1.25rem;
-    fill: currentColor;
-  }
-}
