+++ playbook/app/pb_kits/playbook/pb_rich_text_editor/_rich_text_editor.scss
@@ -9,8 +9,6 @@
 @import "../pb_icon/icon";
 @import "./trix_styles";
 @import "./tiptap_styles";
-@import "./quill_styles";
-@import "./lexxy_styles";
 
 .pb_rich_text_editor_kit {
   // Disables File Attatchment Button in Toolbar
