+++ playbook/app/pb_kits/playbook/pb_rich_text_editor/docs/example.yml
@@ -1,9 +1,7 @@
 examples:
 
   rails:
-  - rich_text_editor_tiptap_default: "TipTap (Vanilla JS)"
-  - rich_text_editor_quill_default: "Quill"
-  - rich_text_editor_lexxy_default: "Lexxy"
+  - rich_text_editor_rails_default: "Rails (TipTap)"
 
   react:
   - rich_text_editor_advanced_default: Advanced Default
