+++ playbook/app/pb_kits/playbook/pb_rich_text_editor/rich_text_editor.html.erb
@@ -1,4 +1,5 @@
-<%# Import map: resolves TipTap/ProseMirror and deps from CDN so one copy loads (avoids "multiple prosemirror-model" error). See README for alternatives. %>
+<%# Import map: single CDN copy of TipTap/ProseMirror (avoids duplicate prosemirror-model). %>
+<%# Sidecar ERB only — not render inline/file — so KitBase#pb_content_tag runs on the component. %>
 <script type="importmap">
   {
     "imports": {
